bugfix for unresponsive UI; report rate change
The tail recursion for reconnect introduced in last commit doesn't seem
to be implemented in chromium and could be the cause of observed
sluggish UI after long running time. It has been replaced with a
setInterval check method.

Reporting frequency of RPM dropped from 50Hz to 10Hz to save overhead.